+++ netwerk/socket/nsNamedPipeService.cpp
@@ -153,7 +153,7 @@ NamedPipeService::AddDataObserver(void* aHandle,
   HANDLE h = CreateIoCompletionPort(aHandle, mIocp,
                                     reinterpret_cast<ULONG_PTR>(aObserver), 1);
   if (NS_WARN_IF(!h)) {
-    LOG_NPS_ERROR("CreateIoCompletionPort error (%d)", GetLastError());
+    LOG_NPS_ERROR("CreateIoCompletionPort error (%lu)", GetLastError());
     return NS_ERROR_FAILURE;
   }
   if (NS_WARN_IF(h != mIocp)) {
@@ -174,7 +174,7 @@ NamedPipeService::AddDataObserver(void* aHandle,
     if (mObservers.Length() == 1) {
       rv = mThread->Dispatch(this, NS_DISPATCH_NORMAL);
       if (NS_WARN_IF(NS_FAILED(rv))) {
-        LOG_NPS_ERROR("Dispatch to thread failed (%08x)", rv);
+        LOG_NPS_ERROR("Dispatch to thread failed (%08x)", uint32_t(rv));
         mObservers.Clear();
         return rv;
       }
@@ -264,7 +264,7 @@ NamedPipeService::Run() {
          * See remarks of |GetQueuedCompletionStatus| API.
          */
 
-        LOG_NPS_ERROR("invalid overlapped (%d)", err);
+        LOG_NPS_ERROR("invalid overlapped (%lu)", err);
         continue;
       }
 
@@ -297,11 +297,11 @@ NamedPipeService::Run() {
     MOZ_ASSERT(obs.get());
 
     if (success) {
-      LOG_NPS_DEBUG("OnDataAvailable: obs=%p, bytes=%d", obs.get(),
+      LOG_NPS_DEBUG("OnDataAvailable: obs=%p, bytes=%lu", obs.get(),
                     bytesTransferred);
       obs->OnDataAvailable(bytesTransferred, overlapped);
     } else {
-      LOG_NPS_ERROR("GetQueuedCompletionStatus %p failed, error=%d", obs.get(),
+      LOG_NPS_ERROR("GetQueuedCompletionStatus %p failed, error=%lu", obs.get(),
                     err);
       obs->OnError(err, overlapped);
     }
